yeet complexity
