--- src/Sentry.Unity.Android/IJniExecutor.cs
@@ -1,10 +1,0 @@
-using System;
-
-namespace Sentry.Unity.Android
-{
-    public interface IJniExecutor : IDisposable
-    {
-        public TResult? Run<TResult>(Func<TResult?> jniOperation);
-        public void Run(Action jniOperation);
-    }
-}
