+++ src/Sentry.Unity.Android/SentryNativeAndroid.cs
@@ -9,8 +9,7 @@ namespace Sentry.Unity.Android
     /// </summary>
     public static class SentryNativeAndroid
     {
-        internal static int? AndroidSdkVersion;
-        internal static IJniExecutor JniExecutor = new JniExecutor();
+        private static readonly JniExecutor JniExecutor = new ();
 
         /// <summary>
         /// Configures the native Android support.
@@ -21,12 +20,6 @@ namespace Sentry.Unity.Android
             if (!options.AndroidNativeSupportEnabled)
             {
                 return;
-            }
-
-            AndroidSdkVersion ??= GetAndroidSdkVersion();
-            if (AndroidSdkVersion is null or <= 23)
-            {
-                JniExecutor = new JniExecutorWithBackgroundWorker();
             }
 
             options.NativeContextWriter = new NativeContextWriter(JniExecutor);
@@ -67,25 +60,6 @@ namespace Sentry.Unity.Android
 
             options.NativeSupportCloseCallback = () => Close(options.DiagnosticLogger);
             options.DefaultUserId = SentryJava.GetInstallationId(JniExecutor);
-        }
-
-        private static int? GetAndroidSdkVersion(IDiagnosticLogger? logger = null)
-        {
-            if (Application.platform == RuntimePlatform.Android)
-            {
-                try
-                {
-                    using var version = new AndroidJavaClass("android.os.Build$VERSION");
-                    return version.GetStatic<int>("SDK_INT");
-                }
-                catch (Exception e)
-                {
-                    logger?.LogError(
-                        e, "Failed to read Android SDK version. Falling back to JNI Execution on background worker.");
-                }
-            }
-
-            return null;
         }
 
         /// <summary>
